Move deferred nonce tests into server.test.ts
Consolidate into the existing integration test file rather than
maintaining a separate file for one feature. Tests are placed
under a new #assignNonce describe block between #signTx and
#broadcastTx to match the transaction lifecycle order.
